+++ src/palm/core/behavior_tree/nodes/leaf/__init__.py
@@ -3,9 +3,11 @@
 from palm.core.behavior_tree.nodes.leaf.action_node import ActionNode
 from palm.core.behavior_tree.nodes.leaf.condition_node import ConditionNode
 from palm.core.behavior_tree.nodes.leaf.interactive_leaf import InteractiveLeaf
+from palm.core.behavior_tree.nodes.leaf.transform_node import TransformLeaf
 
 __all__ = [
     "ActionNode",
     "ConditionNode",
     "InteractiveLeaf",
+    "TransformLeaf",
 ]